Add union types:
One of the original conceptions of microstates was that it should be
an effective way to use modelling your application state using state
machines. One of the key insights was that there are strong analogues
between types and states, where each type corresponds to a distinct
state. So for example, in a light switch state machine, you might have
two types, each descending from a base state:

```javascript

class Switch {
  get isOn() { return false; }
  get isOff() { return false; }

  initialize(value) {
    if (value) {
      return create(On, true);
    } else {
      return create(Off, false);
    }
  }
}

class On extends Switch {
  get isOn() { return true; }

  flip() {
    return create(Off, false);
  }
}

class Off extends Switch {
  get isOff() { return true; }

  flip() {
    return create(On, true);
  }
}
```

This let's you start in one state, and gracefully transition to
another.

```javascript
create(Switch, off).flip().flip().flip() //=> On {}
create(Switch, off).flip().flip().flip().isOn //=> true
```

This is really excellent to use, but expressing it can be fiddly to
get right sometimes. It very much has the feeling of "stitching"
together microstates to achive this nice thing, rather than something
that microstates is expressly there to do.

What we really want is to be able to succinctly express a set of types
and make sure that they adhere together in the following ways:

1. They all descend from the same superclass (On extends Switch, etc..)
2. They have the appropirate boolean state discriminator
   properties (isOn, isOff)
3. The initializers are properly wired together, and contain a
   recursive escape hatch to avoid infinite looping.

This change contains a prototype of a `Union` function that takes a
descriptor and creates a Union type. The union types are based loosely
off of Haskell union types e.g. `data Maybe a = Just a | Nothing`

Where the union type is represented as the superclass, and the
constructors are represented by subclasses. Like Haskell union types,
you the instance you get your hands on _must_ be one of the
subclasses. I.e. I'll never be able to get a reference to a `Maybe`,
instead it will always either be a `Just` or `Nothing`.

The `Union` function returns a class with an initializer that
discriminates between which substate dependending on the value passed
to create.

The `Switch` class above would be rewritten to take advantage of
Unions like so:

```javascript
const Switch = Union({
  On: Switch => class extends Switch {
    flip() {
      return Switch.Off.create();
    }
  },
  Off: Switch => class extends Switch {
    flip() {
      return Switch.On.create();
    }
  }
});
```
Not only is this more concise, but critically, you don't have to
remember to write your boolean discriminators and even more important,
your custom initializers to make sure they coordinate together.

Every Union type has its members as properties of the main union
type. So in our case, `Switch.On` and `Switch.Off`. To create a type
directly, you can use the `create` method on each member.

```javascript
let switch = Switch.Off.create(); Off {}
switch.isOff //=> true
switch.isOn //=> false
switch.type //=> 'Off'

let on = switch.flip(); //=> On {}
switch.isOff //=> false;
switch.isOn //=> true;
switch.type //=> 'On';
````

TODO
====

- Allow adding transitions and methods to the base class.
- Allow for tagged aliases: e.g. create(Color, 'red') instead of
  `create(Color, {type: 'Red'})`
- Allow custom initializers in union members.
- Shorthand for classes that don't have any special transitions